requirements.txt
streamlit>=1.40
pandas>=2.0
scikit-learn>=1.3
openpyxl>=3.1
pypdf>=4.0
Pillow>=10.0
pytesseract>=0.3

